--- a/a5/core/hex.py
+++ b/a5/core/hex.py
@@ -2,10 +2,10 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright (c) A5 contributors
 
-def hex_to_bigint(hex_str: str) -> int:
+def hex_to_bigint(hex_str):
     """Convert hex string to big integer."""
     return int(hex_str, 16)
 
-def bigint_to_hex(index: int) -> str:
+def bigint_to_hex(index):
     """Convert big integer to hex string."""
     return hex(index)[2:]  # Remove '0x' prefix