<!-- Short description of why change is needed -->
#### Background

The Python implementation was surprisingly slow compared the JS. Removing numpy helps as the code is not using it for vectorization much and the overhead slows everything down, see e.g. https://stackoverflow.com/questions/1988091/poor-numpy-cross-performance

Tests using the wireframe example:

## With numpy (main branch)

```
time uv run index.py 4 a5.4.geojson
Successfully generated 3840 A5 cells at resolution 4
Output written to a5.4.geojson

________________________________________________________
Executed in    4.29 secs    fish           external
   usr time    4.43 secs    0.41 millis    4.42 secs
   sys time    0.67 secs    7.30 millis    0.66 secs
```

## Without numpy (this PR)

```
time uv run index.py 4 a5.4.geojson
Successfully generated 3840 A5 cells at resolution 4
Output written to a5.4.geojson

________________________________________________________
Executed in  846.35 millis    fish           external
   usr time  774.47 millis    0.44 millis  774.02 millis
   sys time   49.92 millis    9.14 millis   40.78 millis
```

Perhaps numpy can be added back later for operations which benefit from vectorization. But using it everywhere seems like an expensive antipattern

<!-- For all the PRs -->
#### Change List
-